refactor: 파일명에서 types 제거
