+++ frontend/src/features/auth/contexts/AuthContext.tsx
@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-import type { AuthContextValue } from "@/features/auth/types/auth.types";
+import type { AuthContextValue } from "@/features/auth/types/auth";
 
 export const AuthContext = createContext<AuthContextValue | undefined>(undefined);
